+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.0.0](///compare/v1.1.1...v2.0.0) (2021-11-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* **server-middleware:** 增加了后台 API 调用mongoose 存取数据的功能
+
+### Features
+
+* **server-middleware:** 增加API访问能力 b334e46
+
 ### [1.1.1](///compare/v1.1.0...v1.1.1) (2021-11-15)